fix: don’t limit v3 extensions to default imports
The new approach checks if there is a default export and only then
limits extensions to default import. For the time being, this means
React‘s helpers like useContext() must *still* be accessed via default
React import (e.g., `React.useContext()`), but it’s not required for
most other packages (e.g., with Blueprint it’s not required to use the
awkward `BPCore.ButtonGroup` and the like).